+++ tf.go
@@ -4,8 +4,11 @@ import (
 	"context"
 	"fmt"
 	"io"
+	"os"
+	"os/exec"
 	"runtime"
 	"strings"
+	"time"
 )
 
 const skipCount = 3
@@ -20,7 +23,7 @@ const skipCount = 3
 type TF struct {
 	ctx     context.Context
 	name    string
-	writer  io.Writer
+	output  io.Writer
 	failed  bool
 	skipped bool
 }
@@ -37,7 +40,18 @@ func (tf *TF) Name() string {
 
 // Output returns the io.Writer used to print output.
 func (tf *TF) Output() io.Writer {
-	return tf.writer
+	return tf.output
+}
+
+// Cmd is like exec.Command, but it assigns tf's context
+// and assigns Stdout and Stderr to tf's output,
+// and Stdin to os.Stdin.
+func (tf *TF) Cmd(name string, args ...string) *exec.Cmd {
+	cmd := exec.CommandContext(tf.Context(), name, args...)
+	cmd.Stdin = os.Stdin
+	cmd.Stderr = tf.output
+	cmd.Stdout = tf.output
+	return cmd
 }
 
 // Log formats its arguments using default formatting, analogous to Println,
@@ -125,23 +139,56 @@ func (tf *TF) SkipNow() {
 	runtime.Goexit()
 }
 
+// runResult contains the results of a Action run.
+type runResult struct {
+	Failed   bool
+	Skipped  bool
+	Duration time.Duration
+}
+
+// run executes the action in a seperate goroutine to enable
+// interuption using runtime.Goexit().
+func (tf *TF) run(action func(tf *TF)) runResult {
+	finished := make(chan runResult, 1)
+	go func() {
+		from := time.Now()
+		defer func() {
+			if r := recover(); r != nil {
+				txt := fmt.Sprintf("panic: %v", r)
+				const skipUntilPanic = 3
+				txt = decorate(txt, skipUntilPanic)
+				io.WriteString(tf.output, txt) //nolint // not checking errors when writing to output
+				tf.failed = true
+			}
+			result := runResult{
+				Failed:   tf.failed,
+				Skipped:  tf.skipped,
+				Duration: time.Since(from),
+			}
+			finished <- result
+		}()
+		action(tf)
+	}()
+	return <-finished
+}
+
 // log is used internally in order to provide proper prefix.
 func (tf *TF) log(args ...interface{}) {
 	txt := fmt.Sprint(args...)
-	txt = tf.decorate(txt, skipCount)
-	io.WriteString(tf.writer, txt) //nolint // not checking errors when writing to output
+	txt = decorate(txt, skipCount)
+	io.WriteString(tf.output, txt) //nolint // not checking errors when writing to output
 }
 
 // lof is used internally in order to provide proper prefix.
 func (tf *TF) logf(format string, args ...interface{}) {
 	txt := fmt.Sprintf(format, args...)
-	txt = tf.decorate(txt, skipCount)
-	io.WriteString(tf.writer, txt) //nolint // not checking errors when writing to output
+	txt = decorate(txt, skipCount)
+	io.WriteString(tf.output, txt) //nolint // not checking errors when writing to output
 }
 
 // decorate prefixes the string with the file and line of the call site
 // and inserts the final newline if needed and indentation spaces for formatting.
-func (tf *TF) decorate(s string, skip int) string {
+func decorate(s string, skip int) string {
 	_, file, line, _ := runtime.Caller(skip)
 	if file != "" {
 		// Truncate file name at last file name separator.
